+++ sdk/monitor/opentelemetry-exporter-azuremonitor/tests/trace/test_trace.py
@@ -57,15 +57,16 @@ class TestAzureSpanExporter(unittest.TestCase):
         cls._exporter = AzureMonitorSpanExporter(storage_path=STORAGE_PATH)
 
     def setUp(self):
-        for filename in os.listdir(STORAGE_PATH):
-            file_path = os.path.join(STORAGE_PATH, filename)
-            try:
-                if os.path.isfile(file_path) or os.path.islink(file_path):
-                    os.unlink(file_path)
-                elif os.path.isdir(file_path):
-                    shutil.rmtree(file_path)
-            except OSError as e:
-                print("Failed to delete %s. Reason: %s" % (file_path, e))
+        if os.path.exists(STORAGE_PATH):
+            for filename in os.listdir(STORAGE_PATH):
+                file_path = os.path.join(STORAGE_PATH, filename)
+                try:
+                    if os.path.isfile(file_path) or os.path.islink(file_path):
+                        os.unlink(file_path)
+                    elif os.path.isdir(file_path):
+                        shutil.rmtree(file_path)
+                except OSError as e:
+                    print("Failed to delete %s. Reason: %s" % (file_path, e))
 
     def test_constructor(self):
         """Test the constructor."""
